(PC-14474) refactor(Booking): extract QrCode, DefaultBody, EmailSent, NoTicket components from BookingDetailsTicketContent
